[gnmi] Keep COUNTERS_DB migration in one file
