+++ tests/common/plugins/conditional_mark/tests_mark_conditions.yaml
@@ -2964,12 +2964,6 @@ gnmi/test_gnmi_countersdb.py:
     conditions:
       - "'bmc' in topo_type"
 
-gnmi/test_gnmi_countersdb_config_reload.py:
-  skip:
-    reason: "Not available in BMC"
-    conditions:
-      - "'bmc' in topo_type"
-
 gnmi/test_gnmi_smartswitch.py:
   skip:
     reason: 'SmartSwitch gNMI tests require SmartSwitch platform, not applicable to BMC'
