+++ tests/common/plugins/conditional_mark/tests_mark_conditions_vs_t1_multiasic.yaml
@@ -193,11 +193,6 @@ gnmi/test_gnmi_countersdb.py:
     conditions:
     - asic_type in ['vs'] and 't1-8-lag' in topo_name
     reason: This test case either cannot pass or should be skipped on virtual chassis
-gnmi/test_gnmi_countersdb_config_reload.py:
-  skip:
-    conditions:
-    - asic_type in ['vs'] and 't1-8-lag' in topo_name
-    reason: This test case either cannot pass or should be skipped on virtual chassis
 gnmi/test_gnoi_killprocess.py:
   skip:
     conditions:
